requirements.txt
mojito2
ta
requests
pandas
python-dotenv
streamlit
psutil
anthropic